requirements.txt
gradio
sseclient-py
requests